Refactor code to fix application not loading controller bug
