+++ questionbank/src/main/java/com/example/questionbank/DemoApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class QuestionbankApplication {
+public class DemoApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(QuestionbankApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(DemoApplication.class, args);
+    }
 
 }
