--- questionbank/src/main/java/questionbank/QuestionNotFoundException.java
@@ -1,8 +1,0 @@
-package questionbank;
-
-class QuestionNotFoundException extends RuntimeException {
-
-    QuestionNotFoundException(Long id) {
-        super("Could not find question " + id);
-    }
-}
