--- questionbank/src/main/java/questionbank/QuestionRepository.java
@@ -1,7 +1,0 @@
-package questionbank;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface QuestionRepository extends JpaRepository<Question, Long> {
-
-}
