+++ questionbank/src/main/java/questionbank/commons/QuestionNotFoundAdvice.java
@@ -1,4 +1,4 @@
-package questionbank;
+package questionbank.commons;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
