+++ questionbank/src/main/java/questionbank/commons/QuestionNotFoundException.java
@@ -1,0 +1,8 @@
+package questionbank.commons;
+
+public class QuestionNotFoundException extends RuntimeException {
+
+    public QuestionNotFoundException(Long id) {
+        super("Could not find question " + id);
+    }
+}
