+++ questionbank/src/main/java/questionbank/database/LoadDatabase.java
@@ -1,10 +1,11 @@
-package questionbank;
+package questionbank.database;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import questionbank.model.Question;
 
 @Configuration
 class LoadDatabase {
