+++ questionbank/src/main/java/questionbank/database/QuestionRepository.java
@@ -1,0 +1,8 @@
+package questionbank.database;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import questionbank.model.Question;
+
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+
+}
